Remove unnecessary loop in trigger()
In commit 9ddc9f8 (Simplify trigger code, 2019-12-14), the trigger()
function was simplified by removing an unneeded variable. On second
thought, the whole loop seems unnecessary and can be replaced it with a
simple dictionary lookup.

While at it, also add a log error message when bailing out on an invalid
phase name.