Add argument to disable all Goroutines
When the filter is run with the -disableConcurrency argument, all
requests are processed sequentially. In particular, new filter requests
will not be processed until all previous request have been handled.
Potential delays will accumulate. This option is intended to be used for
testing and debugging only, where immediate and deterministic replies
are helpful. Do not enable this in production.