Implement constant time whitelist queries
With the current subnet whitelist implementation, each lookup requires a
linear scan of the entire whitelist. If the whitelist contains millions
of entries, this can slow down connections significantly.

In other projects, CIDR lookups are often implemented using prefix
trees. This makes lookups very fast but the implementation is
significantly more complex.

An alternative, much simpler solution is to store the whitelisted
subnets in a hash set. To perform a query, one can then apply all 33
possible CIDR masks to the IP address and check whether any of the
resulting CIDR blocks exists in the hash set. With this implementation,
the time complexity of single lookups is still in O(1).

An additional speedup is achieved by maintaining a hash set of the CIDR
masks occurring in the whitelist, instead of always testing all 33
values.

Also, since a strict separation between IP address whitelists and subnet
whitelists are no longer needed, the two features are merged.