Change delay formula
Replace the old formula d - (d / 100) * s by d * (100 - s) / 100 (where
d denotes the slow factor and s denotes the score).

The new formula does not only match to the expected behavior "interpret
the inverse score as a percentage and take that fraction of the slow
factor" more closely, it is also less prone to unwanted rounding
effects. For example, if the score is 100, the new formula is guaranteed
to yield a delay of 0, while the original formula might not.