Mimic DNS queries in the test suite
The current test suite implementation does not only rely on network
connectivity, it also hardcodes real-life IP addresses and assumes they
have a Sender Score within a specific range. This makes the test suite
unusable offline and prone to break with future changes in the score of
those specific IP addresses.

Address this by adding a -testMode command line flag. When used, instead
of executing DNS queries, the Sender Score is derived directly from the
last byte of the connecting IP address. If the IP address ends with 255,
the filter behaves as if the Sender Score DNS query was not found in
production mode. Moreover, if this flag is used, all requests are
processed sequentially.

This patch also removes the -disableConcurrency flag, since sequential
processing is enabled implicitly with -testMode and not useful outside
the test suite.